+++ app/build.gradle.kts
@@ -37,7 +37,6 @@ android {
 
     buildFeatures{
         viewBinding = true
-        dataBinding = true
     }
 }
