--- app/src/main/java/com/example/playlistmaker/api/itunessearch/ItunesSearchResponse.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.api.itunessearch
-
-import com.example.playlistmaker.presentation.search.Track
-
-class ItunesSearchResponse(val results: ArrayList<Track>)
