+++ app/src/main/java/com/example/playlistmaker/data/dto/PlayerRequest.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.data.dto
+
+data class PlayerRequest(val previewUrl: String)
