+++ app/src/main/java/com/example/playlistmaker/data/dto/PlayerResponse.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.data.dto
+
+sealed interface PlayerResponse<T>{
+    data class Data<T>(val value: T): PlayerResponse<T>
+}
