+++ app/src/main/java/com/example/playlistmaker/data/dto/PlayerStatusDto.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.data.dto
+
+enum class PlayerStatusDto {
+    DEFAULT,
+    PREPARED,
+    PLAYING,
+    PAUSED
+}
