+++ app/src/main/java/com/example/playlistmaker/data/dto/TracksSearchRequest.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.data.dto
+
+data class TracksSearchRequest(val expression: String)
