+++ app/src/main/java/com/example/playlistmaker/data/dto/TracksSearchResponse.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.data.dto
+
+data class TracksSearchResponse(val results: List<TrackDto>) : NetworkResponse()
