+++ app/src/main/java/com/example/playlistmaker/data/network/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.data.network
+
+import com.example.playlistmaker.data.dto.NetworkResponse
+
+interface NetworkClient {
+    fun doRequest(dto: Any): NetworkResponse
+}
