+++ app/src/main/java/com/example/playlistmaker/domain/api/TracksRepository.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.domain.api
+
+import com.example.playlistmaker.domain.entity.Resource
+import com.example.playlistmaker.domain.entity.Track
+
+interface TracksRepository {
+    fun searchTracks(expression: String): Resource<ArrayList<Track>>
+}
