+++ app/src/main/java/com/example/playlistmaker/domain/consumer/Consumer.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.domain.consumer
+
+interface Consumer<T> {
+    fun consume(data: ConsumerData<T>)
+}
