+++ app/src/main/java/com/example/playlistmaker/domain/consumer/ConsumerData.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.domain.consumer
+
+sealed interface ConsumerData<T> {
+    data class Data<T>(val value: T) : ConsumerData<T>
+    data class Error<T>(val message: String) : ConsumerData<T>
+}
