+++ app/src/main/java/com/example/playlistmaker/domain/entity/PlayerDetails.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.domain.entity
+
+data class PlayerDetails(
+    val track: Track,
+    val status: PlayerStatus
+)
