+++ app/src/main/java/com/example/playlistmaker/domain/entity/PlayerStatus.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.domain.entity
+
+enum class PlayerStatus {
+    DEFAULT,
+    PREPARED,
+    PLAYING,
+    PAUSED
+}
