+++ app/src/main/java/com/example/playlistmaker/domain/entity/Resource.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.domain.entity
+
+sealed interface Resource<T> {
+    data class Success<T>(val data: T) : Resource<T>
+    data class Error<T>(val message: String) : Resource<T>
+}
