+++ app/src/main/java/com/example/playlistmaker/domain/repository/TrackDetailRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.repository
+
+import com.example.playlistmaker.domain.entity.Track
+
+interface TrackDetailRepository {
+    fun getTrackByString(trackStr: String?): Track?
+}
