--- app/src/main/java/com/example/playlistmaker/presentation/player/PlayerState.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.presentation.player
-
-enum class PlayerState {
-    DEFAULT,
-    PREPARED,
-    PLAYING,
-    PAUSED
-}
