--- app/src/main/java/com/example/playlistmaker/presentation/search/SearchViewGroup.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker.presentation.search
-
-// Настройки отображения сообщения-заглушки в зависимости от возвращаемого ответа
-enum class SearchViewGroup {
-    FIND,
-    HISTORY,
-    PROGRESS,
-    ERROR,
-    FAILURE;
-}
-
