+++ app/src/main/java/com/example/playlistmaker/presentation/tracks/SearchViewGroup.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.presentation.tracks
+
+// Настройки отображения сообщения-заглушки в зависимости от возвращаемого ответа
+enum class SearchViewGroup {
+    FIND,
+    HISTORY,
+    PROGRESS,
+    ERROR,
+    FAILURE;
+}
+
