+++ app/src/main/res/layout/activity_media.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MediaActivity">
+    tools:context=".ui.media.MediaActivity">
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
